Worked on player behaviors and speed reg

+++ Assets/Scripts/Utilities/GaussHouse.cs
@@ -9,16 +9,16 @@ public class GaussHouse : MonoBehaviour
     private static System.Random _rand = new System.Random();
 
     //Uses Box-Muller Transform-https://stackoverflow.com/questions/218060/random-gaussian-variables
-    public static double generateGaussianFloat (float x,float std)
+    public static float generateGaussianFloat (float x,float std)
     {
 
-        
+
         double u1 = 1.0 - _rand.NextDouble();
         double u2 = 1.0 - _rand.NextDouble();
         double randStdNormal = Math.Sqrt(-2.0 * Math.Log(u1)) *
             Math.Sin(2.0 * Math.PI * u2);
 
-        return x + std * randStdNormal;
+        return (float)(x + std * randStdNormal);
     }
     
 }
